embedding/pyproject.toml
[project]
name = "embedding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.6.0",
    "qdrant-client[fastembed]>=1.13.3",
    "sentence-transformers>=4.0.2",
    "webvtt-py>=0.5.1",
    "yt-dlp>=2025.3.31",
]
